move ClientIdentity to java-api
