+++ java-api/src/main/java/com/sap/cloud/security/config/ClientCredentials.java
@@ -3,9 +3,7 @@
  * 
  * SPDX-License-Identifier: Apache-2.0
  */
-package com.sap.cloud.security.client;
-
-import com.sap.xsa.security.container.ClientIdentity;
+package com.sap.cloud.security.config;
 
 import javax.annotation.Nonnull;
 import java.util.Objects;
