+++ token-client/src/main/java/com/sap/cloud/security/xsuaa/tokenflows/PasswordTokenFlow.java
@@ -7,7 +7,7 @@ package com.sap.cloud.security.xsuaa.tokenflows;
 
 import com.sap.cloud.security.xsuaa.Assertions;
 import com.sap.cloud.security.xsuaa.client.*;
-import com.sap.xsa.security.container.ClientIdentity;
+import com.sap.cloud.security.config.ClientIdentity;
 
 import javax.annotation.Nonnull;
 import java.util.Map;
