+++ token-client/src/main/java/com/sap/cloud/security/xsuaa/tokenflows/RefreshTokenFlow.java
@@ -5,9 +5,9 @@
  */
 package com.sap.cloud.security.xsuaa.tokenflows;
 
-import com.sap.cloud.security.client.ClientCredentials;
+import com.sap.cloud.security.config.ClientCredentials;
 import com.sap.cloud.security.xsuaa.client.*;
-import com.sap.xsa.security.container.ClientIdentity;
+import com.sap.cloud.security.config.ClientIdentity;
 import com.sap.xsa.security.container.XSTokenRequest;
 
 import static com.sap.cloud.security.xsuaa.Assertions.assertNotNull;
